APP/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0

# Document Processing
pypdf2==3.0.1
python-docx==1.1.0
pytesseract==0.3.10
pdf2image==1.16.3
pillow==10.1.0

# NLP & AI
transformers==4.35.2
torch==2.1.1
sentence-transformers==2.2.2
spacy==3.7.2
langchain==0.0.350
langchain-community==0.0.1
openai==1.3.8
tiktoken==0.5.2

# Vector Database & RAG
chromadb==0.4.18
faiss-cpu==1.7.4
pinecone-client==2.2.4

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0

# Caching
redis==5.0.1
hiredis==2.2.3

# HTTP & Data Fetching
httpx==0.25.2
beautifulsoup4==4.12.2
requests==2.31.0
aiohttp==3.9.1

# Report Generation
reportlab==4.0.7
weasyprint==60.1
jinja2==3.1.2
pandas==2.1.4
openpyxl==3.1.2

# Security & Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
cryptography==41.0.7

# Utilities
python-dateutil==2.8.2
pytz==2023.3
pyyaml==6.0.1
tenacity==8.2.3

# Monitoring & Logging
python-json-logger==2.0.7
prometheus-client==0.19.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2
